\name{sitelocation}
\alias{sitelocation}
\docType{data}
\title{ Grazed site locations}
\description{
  Locations along a southwest-northeast transect through the northeastern United States. This is a synthetic gradient calculated from latitude and longitude.
}
\usage{data(sitelocation)}
\format{
  The format is:
 num [1:50] 12.19 12.19 12.41 12.42  8.41 ...
}
\examples{
data(sitelocation)
## maybe str(sitelocation) ; plot(sitelocation) ...
}
\keyword{datasets}
