\name{ecodist-package}
\alias{ecodist-package}
\alias{ecodist}
\docType{package}
\title{
Dissimilarity-Based Functions for Ecological Analysis
}
\description{
Dissimilarity-based analysis functions including ordination and Mantel test functions, intended for use with spatial and community data.
}
\details{

This package contains well-established dissimilarity-based ecological analyses, such as \code{\link{nmds}} and \code{\link{mantel}}, and experimental/research analyses such as \code{\link{xmantel}}. Helper functions such as \code{\link{crosstab}} and \code{\link{cor2m}} facilitate analysis of community data.

Because many of the analyses are time-consuming, this package includes worked examples that can be loaded using \code{data()}.

\packageIndices{ecodist}
}
\author{ 
Sarah Goslee and Dean Urban 

Maintainer: Sarah Goslee <Sarah.Goslee@ars.usda.gov>
}
\keyword{ package }
